♻️ refactor wizard into architecture-specific files
