+++ lib/user-interface/react-app/src/components/wizard/agent-core-runtime-wizard.tsx
@@ -7,26 +7,20 @@ import { generateClient } from "aws-amplify/api";
 import { useContext, useEffect, useMemo, useState } from "react";
 
 import {
-    Alert,
     Box,
     Button,
     Checkbox,
-    ColumnLayout,
     Container,
     FormField,
     Header,
-    Icon,
     Input,
     Modal,
-    Popover,
     RadioGroup,
     Select,
     SpaceBetween,
-    Table,
     Textarea,
     Wizard,
 } from "@cloudscape-design/components";
-import ReactMarkdown from "react-markdown";
 import { KnowledgeBase, McpServer, RuntimeSummary, Tool } from "../../API";
 import { AppContext } from "../../common/app-context";
 import {
@@ -36,62 +30,15 @@ import {
     listKnowledgeBases as listKnowledgeBasesQuery,
     listRuntimeAgents as listRuntimeAgentsQuery,
 } from "../../graphql/queries";
+import { getSingleAgentSteps, isSingleAgentStepValid } from "./architectures/single-agent-steps";
+import { getSwarmAgentSteps, isSwarmStepValid } from "./architectures/swarm-agent-steps";
 import {
     AgentCoreRuntimeConfiguration,
     ArchitectureType,
     SearchType,
     SwarmConfiguration,
 } from "./types";
-
-// Set of dangerous keys that could lead to prototype pollution
-const DANGEROUS_KEYS = new Set(["__proto__", "constructor", "prototype"]);
-
-// Helper function to check for prototype pollution attacks
-const isSafePropertyKey = (key: string): boolean => {
-    return !DANGEROUS_KEYS.has(key);
-};
-
-// Safe deep property setter that prevents prototype pollution using recursion
-const safeDeepSetRecursive = (
-    obj: Record<string, any>,
-    keys: readonly string[],
-    keyIndex: number,
-    value: any,
-): Record<string, any> => {
-    // Base case: we've reached the final key
-    if (keyIndex === keys.length - 1) {
-        const finalKey = keys[keyIndex];
-        return { ...obj, [finalKey]: value };
-    }
-
-    // Recursive case: need to go deeper
-    const currentKey = keys[keyIndex];
-    const currentValue = Object.prototype.hasOwnProperty.call(obj, currentKey)
-        ? obj[currentKey]
-        : null;
-    const nestedObj =
-        typeof currentValue === "object" && currentValue !== null
-            ? currentValue
-            : Object.create(null);
-
-    return {
-        ...obj,
-        [currentKey]: safeDeepSetRecursive(nestedObj, keys, keyIndex + 1, value),
-    };
-};
-
-// Safe deep property setter that prevents prototype pollution
-const safeDeepSet = <T extends Record<string, any>>(obj: T, path: string, value: any): T => {
-    const keys = path.split(".");
-
-    // Validate all keys upfront
-    if (!keys.every(isSafePropertyKey)) {
-        console.error("Invalid property path detected - potential prototype pollution");
-        return obj;
-    }
-
-    return safeDeepSetRecursive(obj, keys, 0, value) as T;
-};
+import { DANGEROUS_KEYS, STEP_MIN_HEIGHT, safeDeepSet } from "./wizard-utils";
 
 interface AgentCoreRuntimeCreatorWizardProps {
     onSubmit: (config: AgentCoreRuntimeConfiguration) => void;
@@ -128,19 +75,23 @@ export default function AgentCoreRuntimeCreatorWizard({
     const [architectureType, setArchitectureType] = useState<ArchitectureType>(
         initialData?.architectureType || "SINGLE",
     );
-    const [swarmConfig, setSwarmConfig] = useState<SwarmConfiguration>({
-        agents: [],
-        agentReferences: [],
-        entryAgent: "",
-        orchestrator: {
-            maxHandoffs: 15,
-            maxIterations: 50,
-            executionTimeoutSeconds: 300,
-            nodeTimeoutSeconds: 60,
+    const [swarmConfig, setSwarmConfig] = useState<SwarmConfiguration>(
+        initialData?.swarmConfig || {
+            agents: [],
+            agentReferences: [],
+            entryAgent: "",
+            orchestrator: {
+                maxHandoffs: 15,
+                maxIterations: 50,
+                executionTimeoutSeconds: 300,
+                nodeTimeoutSeconds: 60,
+            },
+            conversationManager: "sliding_window",
         },
-        conversationManager: "sliding_window",
-    });
+    );
 
+    // When creating a new version, skip the architecture selector step
+    const isNewVersion = !!initialData?.architectureType;
 
     // Initialize enableReranking and enableSearchType from initialData on mount
     useEffect(() => {
@@ -151,32 +102,24 @@ export default function AgentCoreRuntimeCreatorWizard({
             Object.keys(initialData.toolParameters).forEach((toolName) => {
                 if (toolName.startsWith("retrieve_from_kb_")) {
                     const toolParams = initialData.toolParameters![toolName];
-
-                    // Check for reranking configuration
-                    const hasRerankingConfig =
-                        toolParams?.retrieval_cfg?.vectorSearchConfiguration?.rerankingConfiguration
-                            ?.bedrockRerankingConfiguration?.numberOfResults;
-                    if (hasRerankingConfig) {
+                    if (
+                        toolParams?.retrieval_cfg?.vectorSearchConfiguration
+                            ?.rerankingConfiguration?.bedrockRerankingConfiguration?.numberOfResults
+                    ) {
                         rerankingStates[toolName] = true;
                     }
-
-                    // Check for search type configuration
-                    const hasOverrideSearchType =
-                        toolParams?.retrieval_cfg?.vectorSearchConfiguration?.overrideSearchType;
-                    if (hasOverrideSearchType) {
+                    if (
+                        toolParams?.retrieval_cfg?.vectorSearchConfiguration?.overrideSearchType
+                    ) {
                         searchTypeStates[toolName] = true;
                     }
                 }
             });
 
-            if (Object.keys(rerankingStates).length > 0) {
-                setEnableReranking(rerankingStates);
-            }
-            if (Object.keys(searchTypeStates).length > 0) {
-                setEnableSearchType(searchTypeStates);
-            }
+            if (Object.keys(rerankingStates).length > 0) setEnableReranking(rerankingStates);
+            if (Object.keys(searchTypeStates).length > 0) setEnableSearchType(searchTypeStates);
         }
-    }, []); // Run once on mount
+    }, [initialData]);
 
     const [config, setConfig] = useState<AgentCoreRuntimeConfiguration>({
         agentName: initialData?.agentName || "",
@@ -188,82 +131,64 @@ export default function AgentCoreRuntimeCreatorWizard({
         useMemory: initialData?.useMemory || false,
         modelInferenceParameters: initialData?.modelInferenceParameters || {
             modelId: "",
-            parameters: {
-                temperature: 0.2,
-                maxTokens: 3000,
-            },
+            parameters: { temperature: 0.2, maxTokens: 3000 },
         },
     });
 
-    const conversationManagerOptions = [
-        { label: "Sliding Window", value: "sliding_window" },
-        { label: "Summarizing", value: "summarizing" },
-        { label: "None", value: "null" },
-    ];
-
     const apiClient = useMemo(() => generateClient(), []);
 
+    // ----------------------------------------------------------------
+    // Data fetching
+    // ----------------------------------------------------------------
     useEffect(() => {
-        // Guard: wait until appConfig is loaded
         if (!appConfig) return;
-
         let isCancelled = false;
 
         const fetchData = async () => {
             try {
                 const knowledgeBaseSupported = appConfig?.knowledgeBaseIsSupported ?? false;
-
-                const [kbResult, toolsResult, mcpServersResult, agentsResult] = await Promise.all([
-                    knowledgeBaseSupported
-                        ? apiClient.graphql({ query: listKnowledgeBasesQuery })
-                        : Promise.resolve({ data: { listKnowledgeBases: [] } }),
-                    apiClient.graphql({ query: listAvailableToolsQuery }),
-                    apiClient.graphql({ query: listAvailableMcpServersQuery }),
-                    apiClient.graphql({ query: listRuntimeAgentsQuery }),
-                ]);
+                const [kbResult, toolsResult, mcpServersResult, agentsResult] =
+                    await Promise.all([
+                        knowledgeBaseSupported
+                            ? apiClient.graphql({ query: listKnowledgeBasesQuery })
+                            : Promise.resolve({ data: { listKnowledgeBases: [] } }),
+                        apiClient.graphql({ query: listAvailableToolsQuery }),
+                        apiClient.graphql({ query: listAvailableMcpServersQuery }),
+                        apiClient.graphql({ query: listRuntimeAgentsQuery }),
+                    ]);
 
                 if (isCancelled) return;
-
                 setKnowledgeBases(kbResult.data!.listKnowledgeBases);
-                if (toolsResult.data?.listAvailableTools) {
+                if (toolsResult.data?.listAvailableTools)
                     setAvailableTools(toolsResult.data.listAvailableTools);
-                }
-                if (mcpServersResult.data?.listAvailableMcpServers) {
+                if (mcpServersResult.data?.listAvailableMcpServers)
                     setAvailableMcpServers(mcpServersResult.data.listAvailableMcpServers);
-                }
-                if (agentsResult.data?.listRuntimeAgents) {
+                if (agentsResult.data?.listRuntimeAgents)
                     setAvailableAgents(agentsResult.data.listRuntimeAgents);
-                }
             } catch (error) {
                 console.error("Failed to fetch data:", error);
             }
         };
         fetchData();
-
-        return () => {
-            isCancelled = true;
-        };
+        return () => { isCancelled = true; };
     }, [appConfig, apiClient]);
 
     useEffect(() => {
         if (appConfig && appConfig.aws_bedrock_supported_models) {
             const models = Object.entries(appConfig.aws_bedrock_supported_models).map(
-                ([label, value]) => {
-                    const modelValue = value.replace(
-                        "[REGION-PREFIX]",
-                        appConfig.aws_project_region.split("-")[0],
-                    );
-                    return { label, value: modelValue };
-                },
+                ([label, value]) => ({
+                    label,
+                    value: value.replace("[REGION-PREFIX]", appConfig.aws_project_region.split("-")[0]),
+                }),
             );
             setModelOptions(models);
 
             if (!config.modelInferenceParameters.modelId) {
                 const defaultModel = models.find(
-                    (model) =>
-                        model.label.toLowerCase().includes("claude") &&
-                        model.label.toLowerCase().includes("haiku") &&
-                        model.label.toLowerCase().includes("4.5"),
+                    (m) =>
+                        m.label.toLowerCase().includes("claude") &&
+                        m.label.toLowerCase().includes("haiku") &&
+                        m.label.toLowerCase().includes("4.5"),
                 );
                 if (defaultModel) {
                     setConfig((prev) => ({
@@ -276,62 +201,44 @@ export default function AgentCoreRuntimeCreatorWizard({
                 }
             }
 
-            // Set up reranking models from configuration
-            if (appConfig && appConfig.aws_bedrock_supported_reranking_models) {
-                console.log(
-                    "Reranking models config:",
-                    appConfig.aws_bedrock_supported_reranking_models,
-                );
-                const rerankingModels = Object.entries(
-                    appConfig.aws_bedrock_supported_reranking_models,
-                ).map(([label, value]) => {
-                    return { label, value: value };
-                });
-                console.log("Processed reranking models:", rerankingModels);
-                setRerankingModelOptions(rerankingModels);
-            } else {
-                console.log(
-                    "No reranking models found in appConfig:",
-                    appConfig?.aws_bedrock_supported_reranking_models,
+            if (appConfig.aws_bedrock_supported_reranking_models) {
+                setRerankingModelOptions(
+                    Object.entries(appConfig.aws_bedrock_supported_reranking_models).map(
+                        ([label, value]) => ({ label, value }),
+                    ),
                 );
             }
         }
     }, [appConfig, config.modelInferenceParameters.modelId]);
 
+    // ----------------------------------------------------------------
+    // Tool / KB / MCP / Sub-agent actions (used by single-agent steps)
+    // ----------------------------------------------------------------
     const addTool = (toolName: string | undefined) => {
         if (!toolName || toolName === "retrieve_from_kb" || config.tools.includes(toolName)) return;
-
         setConfig((prev) => ({
             ...prev,
             tools: [...prev.tools, toolName],
-            toolParameters: {
-                ...prev.toolParameters,
-                [toolName]: {},
-            },
+            toolParameters: { ...prev.toolParameters, [toolName]: {} },
         }));
     };
 
-    const addMcpServer = (serverName: string | undefined) => {
-        if (!serverName || config.mcpServers.includes(serverName)) return;
-
-        setConfig((prev) => ({
-            ...prev,
-            mcpServers: [...prev.mcpServers, serverName],
-        }));
-    };
-
-    const removeMcpServer = (serverName: string) => {
-        setConfig((prev) => ({
-            ...prev,
-            mcpServers: prev.mcpServers.filter((server) => server !== serverName),
-        }));
+    const removeTool = (toolName: string) => {
+        setConfig((prev) => {
+            const newToolParameters = { ...prev.toolParameters };
+            delete newToolParameters[toolName];
+            return {
+                ...prev,
+                tools: prev.tools.filter((t) => t !== toolName),
+                toolParameters: newToolParameters,
+            };
+        });
     };
 
     const addSubAgent = (agentName: string | undefined) => {
         if (!agentName) return;
         const toolName = `invoke_subagent_${agentName}`;
         if (config.tools.includes(toolName)) return;
-
         setConfig((prev) => ({
             ...prev,
             tools: [...prev.tools, toolName],
@@ -342,33 +249,29 @@ export default function AgentCoreRuntimeCreatorWizard({
         }));
     };
 
-    const removeTool = (toolName: string) => {
-        setConfig((prev) => {
-            const newToolParameters = { ...prev.toolParameters };
-            delete newToolParameters[toolName];
-            return {
-                ...prev,
-                tools: prev.tools.filter((tool) => tool !== toolName),
-                toolParameters: newToolParameters,
-            };
-        });
+    const addMcpServer = (serverName: string | undefined) => {
+        if (!serverName || config.mcpServers.includes(serverName)) return;
+        setConfig((prev) => ({ ...prev, mcpServers: [...prev.mcpServers, serverName] }));
+    };
+
+    const removeMcpServer = (serverName: string) => {
+        setConfig((prev) => ({
+            ...prev,
+            mcpServers: prev.mcpServers.filter((s) => s !== serverName),
+        }));
     };
 
     const addKnowledgeBase = (kbId: string | undefined) => {
         if (!kbId) return;
         const toolName = `retrieve_from_kb_${kbId}`;
-
+        if (config.tools.includes(toolName)) return;
         setConfig((prev) => ({
             ...prev,
             tools: [...prev.tools, toolName],
             toolParameters: {
                 ...prev.toolParameters,
                 [toolName]: {
-                    retrieval_cfg: {
-                        vectorSearchConfiguration: {
-                            numberOfResults: "5",
-                        },
-                    },
+                    retrieval_cfg: { vectorSearchConfiguration: { numberOfResults: "5" } },
                     kb_id: kbId,
                 },
             },
@@ -377,27 +280,16 @@ export default function AgentCoreRuntimeCreatorWizard({
 
     const updateToolParameter = (toolName: string, paramPath: string, value: any) => {
         setConfig((prev) => {
-            // Validate toolName to prevent prototype pollution
             if (DANGEROUS_KEYS.has(toolName)) {
                 console.error("Invalid tool name detected - potential prototype pollution");
                 return prev;
             }
-
             const currentToolParams = prev.toolParameters[toolName] || {};
             const updatedToolParams = safeDeepSet(currentToolParams, paramPath, value);
-
-            // safeDeepSet always creates a new object on success via spread operator.
-            // Same reference means validation failed (dangerous key detected).
-            if (updatedToolParams === currentToolParams) {
-                return prev;
-            }
-
+            if (updatedToolParams === currentToolParams) return prev;
             return {
                 ...prev,
-                toolParameters: {
-                    ...prev.toolParameters,
-                    [toolName]: updatedToolParams,
-                },
+                toolParameters: { ...prev.toolParameters, [toolName]: updatedToolParams },
             };
         });
     };
@@ -405,26 +297,14 @@ export default function AgentCoreRuntimeCreatorWizard({
     const openConfigureModal = async (toolName: string) => {
         if (toolName.startsWith("retrieve_from_kb_")) {
             setSelectedKbForConfig(toolName);
-
-            // Check if overrideSearchType exists in config and update enableSearchType
             const hasOverrideSearchType =
                 config.toolParameters[toolName]?.retrieval_cfg?.vectorSearchConfiguration
                     ?.overrideSearchType;
-
-            setEnableSearchType((prev) => ({
-                ...prev,
-                [toolName]: !!hasOverrideSearchType,
-            }));
-
-            // Check if reranking configuration exists and update enableReranking
+            setEnableSearchType((prev) => ({ ...prev, [toolName]: !!hasOverrideSearchType }));
             const hasRerankingConfig =
                 config.toolParameters[toolName]?.retrieval_cfg?.vectorSearchConfiguration
                     ?.rerankingConfiguration?.bedrockRerankingConfiguration?.numberOfResults;
-
-            setEnableReranking((prev) => ({
-                ...prev,
-                [toolName]: !!hasRerankingConfig,
-            }));
+            setEnableReranking((prev) => ({ ...prev, [toolName]: !!hasRerankingConfig }));
         } else {
             setSelectedToolForConfig(toolName);
             if (toolName.startsWith("invoke_subagent_")) {
@@ -457,6 +337,9 @@ export default function AgentCoreRuntimeCreatorWizard({
         setSelectedToolForConfig(null);
     };
 
+    // ----------------------------------------------------------------
+    // Swarm-specific actions
+    // ----------------------------------------------------------------
     const addAgentReference = (agentName: string) => {
         if (swarmConfig.agentReferences.some((r) => r.agentName === agentName)) return;
         setSwarmConfig((prev) => ({
@@ -488,58 +371,18 @@ export default function AgentCoreRuntimeCreatorWizard({
         return swarmConfig.agentReferences.map((r) => r.agentName);
     };
 
-    const isStepValid = (stepIndex: number) => {
-        // Step 0 is always Architecture Type - always valid
-        if (stepIndex === 0) return true;
-
-        if (architectureType === "SINGLE") {
-            // Step 1 = Basic Config
-            if (stepIndex === 1) {
-                const agentNamePattern = /^[a-zA-Z][a-zA-Z0-9_]{0,47}$/;
-                return (
-                    config.instructions.trim() !== "" &&
-                    config.agentName.trim() !== "" &&
-                    agentNamePattern.test(config.agentName)
-                );
-            }
-            // Step 4 = Tools Config (index shifts by 1 due to architecture step)
-            if (stepIndex === 4) {
-                return !config.tools.some((tool) => {
-                    return (
-                        tool.startsWith("invoke_subagent_") &&
-                        !config.toolParameters[tool]?.agentName
-                    );
-                });
-            }
-        } else {
-            // SWARM: Step 1 = Swarm Configuration (agent name + agents + entry agent)
-            if (stepIndex === 1) {
-                const agentNamePattern = /^[a-zA-Z][a-zA-Z0-9_]{0,47}$/;
-                const hasAgentName =
-                    config.agentName.trim() !== "" &&
-                    agentNamePattern.test(config.agentName);
-                const hasAgents = swarmConfig.agentReferences.length > 0;
-                const hasEntryAgent = swarmConfig.entryAgent.trim() !== "";
-                return hasAgentName && hasAgents && hasEntryAgent;
-            }
-        }
-        return true;
-    };
+    // ----------------------------------------------------------------
+    // Derived data for single-agent steps
+    // ----------------------------------------------------------------
+    const knowledgeBaseIsSupported = appConfig?.knowledgeBaseIsSupported ?? false;
 
     const availableKnowledgeBases = knowledgeBases
         .filter((kb) => !config.tools.some((tool) => tool === `retrieve_from_kb_${kb.id}`))
-        .map((kb: KnowledgeBase) => ({
-            label: kb.description || kb.name,
-            value: kb.id,
-        }));
+        .map((kb) => ({ label: kb.description || kb.name, value: kb.id }));
 
     const availableToolsOptions = availableTools
         .filter((tool) => !tool.invokesSubAgent && !config.tools.includes(tool.name))
-        .map((tool) => ({
-            label: tool.name,
-            value: tool.name,
-            description: tool.description,
-        }));
+        .map((tool) => ({ label: tool.name, value: tool.name, description: tool.description || undefined }));
 
     const availableSubAgents = availableAgents
         .filter(
@@ -547,55 +390,38 @@ export default function AgentCoreRuntimeCreatorWizard({
                 agent.agentName !== config.agentName &&
                 !config.tools.includes(`invoke_subagent_${agent.agentName}`),
         )
-        .map((agent) => ({
-            label: agent.agentName,
-            value: agent.agentName,
-        }));
+        .map((agent) => ({ label: agent.agentName, value: agent.agentName }));
 
     const availableMcpServersOptions = availableMcpServers
-        .filter((mcpServer) => !config.mcpServers.includes(mcpServer.name))
-        .map((mcpServer) => ({
-            label: mcpServer.name,
-            value: mcpServer.name,
-            description: mcpServer.description,
-        }));
+        .filter((s) => !config.mcpServers.includes(s.name))
+        .map((s) => ({ label: s.name, value: s.name, description: s.description || undefined }));
 
     const selectedMcpServersData = config.mcpServers.map((serverName) => {
         const serverInfo = availableMcpServers.find((s) => s.name === serverName);
         return {
             name: serverName,
             description: serverInfo?.description || "No description available",
-            // identityName: serverInfo?.identityName || "Unknown",
             mcpUrl: serverInfo?.mcpUrl || "",
         };
     });
 
     const selectedToolsData = config.tools
-        .filter(
-            (tool) => !tool.startsWith("retrieve_from_kb_") && !tool.startsWith("invoke_subagent_"),
-        )
+        .filter((t) => !t.startsWith("retrieve_from_kb_") && !t.startsWith("invoke_subagent_"))
         .map((toolName) => {
             const toolInfo = availableTools.find((t) => t.name === toolName);
-            return {
-                name: toolName,
-                description: toolInfo?.description || "No description available",
-            };
+            return { name: toolName, description: toolInfo?.description || "No description available" };
         });
 
     const selectedSubAgentsData = config.tools
-        .filter((tool) => tool.startsWith("invoke_subagent_"))
-        .map((toolName) => {
-            const agentName = toolName.replace("invoke_subagent_", "");
-            const params = config.toolParameters[toolName];
-            return {
-                toolName,
-                agentName,
-                params,
-            };
-        });
+        .filter((t) => t.startsWith("invoke_subagent_"))
+        .map((toolName) => ({
+            toolName,
+            agentName: toolName.replace("invoke_subagent_", ""),
+            params: config.toolParameters[toolName],
+        }));
 
     const selectedKnowledgeBasesData = config.tools
-        .filter((tool) => tool.startsWith("retrieve_from_kb_"))
+        .filter((t) => t.startsWith("retrieve_from_kb_"))
         .map((toolName) => {
             const kbId = toolName.replace("retrieve_from_kb_", "");
             const kb = knowledgeBases.find((k) => k.id === kbId);
@@ -614,986 +440,105 @@ export default function AgentCoreRuntimeCreatorWizard({
         ? config.toolParameters[selectedToolForConfig]
         : null;
 
-    const stepMinHeight = "62vh";
-
-    const knowledgeBaseIsSupported = appConfig?.knowledgeBaseIsSupported ?? false;
-
-    const steps = [
-        {
-            title: "Architecture Type",
-            content: (
-                <div style={{ minHeight: stepMinHeight }}>
-                    <Container header={<Header variant="h2">Select Architecture Type</Header>}>
-                        <SpaceBetween direction="vertical" size="l">
-                            <FormField
-                                label="Architecture"
-                                description="Choose the agent architecture for this runtime"
-                            >
-                                <RadioGroup
-                                    value={architectureType}
-                                    onChange={({ detail }) =>
-                                        setArchitectureType(detail.value as ArchitectureType)
-                                    }
-                                    items={[
-                                        {
-                                            value: "SINGLE",
-                                            label: "Single Agent / Agents as Tools",
-                                            description:
-                                                "A single agent with tools, knowledge bases, MCP servers, and optional sub-agents invoked as tools",
-                                        },
-                                        {
-                                            value: "SWARM",
-                                            label: "Swarm",
-                                            description:
-                                                "Multiple specialized agents that collaborate via handoffs",
-                                        },
-                                    ]}
-                                />
-                            </FormField>
-                        </SpaceBetween>
-                    </Container>
-                </div>
-            ),
-        },
-        ...(architectureType === "SINGLE"
-            ? [
-                  {
-                      title: "Basic Configuration",
-            content: (
-                <div style={{ minHeight: stepMinHeight }}>
-                    <Container header={<Header variant="h2">Agent Instructions</Header>}>
-                        <SpaceBetween direction="vertical" size="l">
-                            <FormField
-                                label="Agent Name"
-                                description="Enter a unique name for your agent"
-                                errorText={
-                                    config.agentName.trim() === ""
-                                        ? "Agent name is required"
-                                        : !/^[a-zA-Z][a-zA-Z0-9_]{0,47}$/.test(config.agentName)
-                                          ? "Agent name must start with a letter and contain only letters, numbers, and underscores (max 48 characters)"
-                                          : ""
+    // ----------------------------------------------------------------
+    // Build steps: architecture selector + architecture-specific steps
+    // ----------------------------------------------------------------
+    const architectureStep = {
+        title: "Architecture Type",
+        content: (
+            <div style={{ minHeight: STEP_MIN_HEIGHT }}>
+                <Container header={<Header variant="h2">Select Architecture Type</Header>}>
+                    <SpaceBetween direction="vertical" size="l">
+                        <FormField
+                            label="Architecture"
+                            description="Choose the agent architecture for this runtime"
+                        >
+                            <RadioGroup
+                                value={architectureType}
+                                onChange={({ detail }) =>
+                                    setArchitectureType(detail.value as ArchitectureType)
                                 }
-                            >
-                                <Input
-                                    value={config.agentName}
-                                    onChange={({ detail }) =>
-                                        setConfig((prev) => ({ ...prev, agentName: detail.value }))
-                                    }
-                                    placeholder="Enter agent name..."
-                                    invalid={config.agentName.trim() === ""}
-                                />
-                            </FormField>
-                            <FormField
-                                label="Instructions"
-                                description="Provide detailed instructions for your agent"
-                                errorText={
-                                    config.instructions.trim() === ""
-                                        ? "Instructions are required"
-                                        : ""
-                                }
-                            >
-                                <Textarea
-                                    value={config.instructions}
-                                    onChange={({ detail }) =>
-                                        setConfig((prev) => ({
-                                            ...prev,
-                                            instructions: detail.value,
-                                        }))
-                                    }
-                                    placeholder="Enter agent instructions..."
-                                    rows={8}
-                                    invalid={config.instructions.trim() === ""}
-                                />
-                            </FormField>
-                            <FormField label="Conversation Manager">
-                                <Select
-                                    selectedOption={
-                                        conversationManagerOptions.find(
-                                            (opt) => opt.value === config.conversationManager,
-                                        ) || null
-                                    }
-                                    onChange={({ detail }) =>
-                                        setConfig((prev) => ({
-                                            ...prev,
-                                            conversationManager: (detail.selectedOption?.value ||
-                                                "sliding_window") as
-                                                | "null"
-                                                | "sliding_window"
-                                                | "summarizing",
-                                        }))
-                                    }
-                                    options={conversationManagerOptions}
-                                />
-                            </FormField>
-                        </SpaceBetween>
-                    </Container>
-                </div>
-            ),
-        },
-        {
-            title: "Model Configuration",
-            content: (
-                <div style={{ minHeight: stepMinHeight }}>
-                    <Container header={<Header variant="h2">Model & Parameters</Header>}>
-                        <SpaceBetween direction="vertical" size="l">
-                            <FormField label="Model">
-                                <Select
-                                    selectedOption={
-                                        modelOptions.find(
-                                            (opt) =>
-                                                opt.value ===
-                                                config.modelInferenceParameters.modelId,
-                                        ) || null
-                                    }
-                                    onChange={({ detail }) =>
-                                        setConfig((prev) => ({
-                                            ...prev,
-                                            modelInferenceParameters: {
-                                                ...prev.modelInferenceParameters,
-                                                modelId: detail.selectedOption?.value || "",
-                                            },
-                                        }))
-                                    }
-                                    options={modelOptions}
-                                />
-                            </FormField>
-                            <FormField label="Temperature" description="Value between 0 and 1">
-                                <Input
-                                    value={config.modelInferenceParameters.parameters.temperature.toString()}
-                                    onChange={({ detail }) => {
-                                        const value = parseFloat(detail.value) || 0;
-                                        if (value >= 0 && value <= 1) {
-                                            setConfig((prev) => ({
-                                                ...prev,
-                                                modelInferenceParameters: {
-                                                    ...prev.modelInferenceParameters,
-                                                    parameters: {
-                                                        ...prev.modelInferenceParameters.parameters,
-                                                        temperature: value,
-                                                    },
-                                                },
-                                            }));
-                                        }
-                                    }}
-                                    type="number"
-                                    step={0.05}
-                                />
-                            </FormField>
-                            <FormField label="Max Tokens" description="Value between 100 and 4000">
-                                <Input
-                                    value={config.modelInferenceParameters.parameters.maxTokens.toString()}
-                                    onChange={({ detail }) => {
-                                        const value = parseInt(detail.value) || 100;
-                                        if (value >= 100 && value <= 4000) {
-                                            setConfig((prev) => ({
-                                                ...prev,
-                                                modelInferenceParameters: {
-                                                    ...prev.modelInferenceParameters,
-                                                    parameters: {
-                                                        ...prev.modelInferenceParameters.parameters,
-                                                        maxTokens: value,
-                                                    },
-                                                },
-                                            }));
-                                        }
-                                    }}
-                                    type="number"
-                                    step={100}
-                                />
-                            </FormField>
-                        </SpaceBetween>
-                    </Container>
-                </div>
-            ),
-        },
-        {
-            title: "Memory Configuration",
-            content: (
-                <div style={{ minHeight: stepMinHeight }}>
-                    <Container header={<Header variant="h2">Memory Settings</Header>}>
-                        <SpaceBetween direction="vertical" size="l">
-                            <FormField
-                                label="AgentCore Memory"
-                                description="Create an AgentCore Memory and attach it to your agent Runtime."
-                            >
-                                <Checkbox
-                                    checked={config.useMemory || false}
-                                    onChange={({ detail }) =>
-                                        setConfig((prev) => ({
-                                            ...prev,
-                                            useMemory: detail.checked,
-                                        }))
-                                    }
-                                >
-                                    Enable AgentCore Memory
-                                </Checkbox>
-                            </FormField>
-                            {config.useMemory && (
-                                <Alert type="info" header="AgentCore Memory Enabled">
-                                    AgentCore Memory will be created and attached to your agent
-                                    Runtime. This allows the agent to maintain conversation context
-                                    even when sessions are terminated (due to inactivity or reaching
-                                    max duration).
-                                </Alert>
-                            )}
-                        </SpaceBetween>
-                    </Container>
-                </div>
-            ),
-        },
-        {
-            title: "Tools Configuration",
-            content: (
-                <div style={{ minHeight: stepMinHeight }}>
-                    <Container header={<Header variant="h2">Configure Tools</Header>}>
-                        <SpaceBetween direction="vertical" size="l">
-                            <FormField label="Available Tools">
-                                <Select
-                                    placeholder={
-                                        availableToolsOptions.length === 0
-                                            ? "No additional tools available"
-                                            : "Select a tool to add"
-                                    }
-                                    options={availableToolsOptions}
-                                    disabled={availableToolsOptions.length === 0}
-                                    onChange={({ detail }) => {
-                                        if (detail.selectedOption) {
-                                            addTool(detail.selectedOption.value);
-                                        }
-                                    }}
-                                    selectedOption={null}
-                                />
-                            </FormField>
+                                items={[
+                                    {
+                                        value: "SINGLE",
+                                        label: "Single Agent / Agents as Tools",
+                                        description:
+                                            "A single agent with tools, knowledge bases, MCP servers, and optional sub-agents invoked as tools",
+                                    },
+                                    {
+                                        value: "SWARM",
+                                        label: "Swarm",
+                                        description:
+                                            "Multiple specialized agents that collaborate via handoffs",
+                                    },
+                                ]}
+                            />
+                        </FormField>
+                    </SpaceBetween>
+                </Container>
+            </div>
+        ),
+    };
 
-                            <FormField
-                                label="Available Sub-Agents"
-                                description="Sub-agents that this orchestrator can invoke as tools"
-                            >
-                                <Select
-                                    placeholder={
-                                        availableSubAgents.length === 0
-                                            ? "No additional sub-agents available"
-                                            : "Select a sub-agent to add as a tool"
-                                    }
-                                    options={availableSubAgents}
-                                    disabled={availableSubAgents.length === 0}
-                                    onChange={({ detail }) => {
-                                        if (detail.selectedOption) {
-                                            addSubAgent(detail.selectedOption.value);
-                                        }
-                                    }}
-                                    selectedOption={null}
-                                />
-                            </FormField>
+    const architectureSpecificSteps =
+        architectureType === "SINGLE"
+            ? getSingleAgentSteps({
+                  config,
+                  setConfig,
+                  modelOptions,
+                  availableToolsOptions,
+                  availableSubAgents,
+                  availableMcpServersOptions,
+                  availableKnowledgeBases,
+                  selectedToolsData,
+                  selectedSubAgentsData,
+                  selectedMcpServersData,
+                  selectedKnowledgeBasesData,
+                  knowledgeBaseIsSupported,
+                  isCreating,
+                  addTool,
+                  removeTool,
+                  addSubAgent,
+                  addMcpServer,
+                  removeMcpServer,
+                  addKnowledgeBase,
+                  openConfigureModal,
+              })
+            : getSwarmAgentSteps({
+                  config,
+                  setConfig,
+                  swarmConfig,
+                  setSwarmConfig,
+                  availableAgents,
+                  isCreating,
+                  architectureType,
+                  addAgentReference,
+                  removeAgentReference,
+                  updateAgentReferenceEndpoint,
+                  getSwarmAgentNames,
+              });
 
-                            <Container header={<Header variant="h3">Selected Tools</Header>}>
-                                {selectedToolsData.length === 0 ? (
-                                    <Alert type="info">No tools selected</Alert>
-                                ) : (
-                                    <Table
-                                        columnDefinitions={[
-                                            {
-                                                id: "name",
-                                                header: "Tool Name",
-                                                cell: (item) => item.name,
-                                            },
-                                            {
-                                                id: "description",
-                                                header: "Description",
-                                                cell: (item) => (
-                                                    <Popover
-                                                        dismissButton={false}
-                                                        position="top"
-                                                        size="medium"
-                                                        triggerType="custom"
-                                                        content={
-                                                            <Box padding="xs">
-                                                                <ReactMarkdown>
-                                                                    {item.description}
-                                                                </ReactMarkdown>
-                                                            </Box>
-                                                        }
-                                                    >
-                                                        <Icon name="status-info" />
-                                                    </Popover>
-                                                ),
-                                            },
-                                            {
-                                                id: "actions",
-                                                header: "Actions",
-                                                cell: (item) => (
-                                                    <Button
-                                                        variant="icon"
-                                                        iconName="close"
-                                                        onClick={() => removeTool(item.name)}
-                                                    />
-                                                ),
-                                            },
-                                        ]}
-                                        items={selectedToolsData}
-                                        loadingText="Loading tools"
-                                        empty={
-                                            <Box textAlign="center" color="inherit">
-                                                <b>No tools selected</b>
-                                                <Box
-                                                    padding={{ bottom: "s" }}
-                                                    variant="p"
-                                                    color="inherit"
-                                                >
-                                                    Select tools from the dropdown above.
-                                                </Box>
-                                            </Box>
-                                        }
-                                    />
-                                )}
-                            </Container>
+    const steps = isNewVersion
+        ? [...architectureSpecificSteps]
+        : [architectureStep, ...architectureSpecificSteps];
 
-                            <Container
-                                header={<Header variant="h3">Selected Sub-Agents (Tools)</Header>}
-                            >
-                                {selectedSubAgentsData.length === 0 ? (
-                                    <Alert type="info">No sub-agents selected</Alert>
-                                ) : (
-                                    <Table
-                                        columnDefinitions={[
-                                            {
-                                                id: "name",
-                                                header: "Agent Name",
-                                                cell: (item) => item.agentName,
-                                            },
-                                            {
-                                                id: "parameters",
-                                                header: "Parameters",
-                                                cell: (item) => {
-                                                    const isConfigured =
-                                                        item.params?.qualifier && item.params?.role;
-                                                    return (
-                                                        <Button
-                                                            variant="normal"
-                                                            onClick={() =>
-                                                                openConfigureModal(item.toolName)
-                                                            }
-                                                        >
-                                                            {isConfigured
-                                                                ? "Configured"
-                                                                : "Configure"}
-                                                        </Button>
-                                                    );
-                                                },
-                                            },
-                                            {
-                                                id: "actions",
-                                                header: "Actions",
-                                                cell: (item) => (
-                                                    <Button
-                                                        variant="icon"
-                                                        iconName="close"
-                                                        onClick={() => removeTool(item.toolName)}
-                                                    />
-                                                ),
-                                            },
-                                        ]}
-                                        items={selectedSubAgentsData}
-                                        loadingText="Loading sub-agents"
-                                        empty={
-                                            <Box textAlign="center" color="inherit">
-                                                <b>No sub-agents selected</b>
-                                                <Box
-                                                    padding={{ bottom: "s" }}
-                                                    variant="p"
-                                                    color="inherit"
-                                                >
-                                                    Select sub-agents from the dropdown above.
-                                                </Box>
-                                            </Box>
-                                        }
-                                    />
-                                )}
-                            </Container>
-                        </SpaceBetween>
-                    </Container>
-                </div>
-            ),
-        },
-        {
-            title: "MCP Servers Configuration",
-            content: (
-                <div style={{ minHeight: stepMinHeight }}>
-                    <Container header={<Header variant="h2">Configure MCP Servers</Header>}>
-                        <SpaceBetween direction="vertical" size="l">
-                            <FormField
-                                label="Available MCP Servers"
-                                description="Model Context Protocol servers provide additional capabilities to your agent"
-                            >
-                                <Select
-                                    placeholder={
-                                        availableMcpServersOptions.length === 0
-                                            ? "No MCP servers available"
-                                            : "Select an MCP server to add"
-                                    }
-                                    options={availableMcpServersOptions}
-                                    disabled={availableMcpServersOptions.length === 0}
-                                    onChange={({ detail }) => {
-                                        if (detail.selectedOption) {
-                                            addMcpServer(detail.selectedOption.value);
-                                        }
-                                    }}
-                                    selectedOption={null}
-                                />
-                            </FormField>
+    const isStepValid = (stepIndex: number) => {
+        if (!isNewVersion) {
+            if (stepIndex === 0) return true; // Architecture type step
+            const archStepIndex = stepIndex - 1;
+            return architectureType === "SINGLE"
+                ? isSingleAgentStepValid(archStepIndex, config)
+                : isSwarmStepValid(archStepIndex, config, swarmConfig);
+        }
+        // New version: no architecture step offset
+        return architectureType === "SINGLE"
+            ? isSingleAgentStepValid(stepIndex, config)
+            : isSwarmStepValid(stepIndex, config, swarmConfig);
+    };
 
-                            <Container header={<Header variant="h3">Selected MCP Servers</Header>}>
-                                {selectedMcpServersData.length === 0 ? (
-                                    <Alert type="info">No MCP servers selected</Alert>
-                                ) : (
-                                    <Table
-                                        columnDefinitions={[
-                                            {
-                                                id: "name",
-                                                header: "Server Name",
-                                                cell: (item) => item.name,
-                                            },
-                                            // {
-                                            //     id: "identityName",
-                                            //     header: "Identity Name",
-                                            //     cell: (item) => item.identityName,
-                                            // },
-                                            {
-                                                id: "description",
-                                                header: "Description",
-                                                cell: (item) => (
-                                                    <Popover
-                                                        dismissButton={false}
-                                                        position="top"
-                                                        size="medium"
-                                                        triggerType="custom"
-                                                        content={
-                                                            <Box padding="xs">
-                                                                <ReactMarkdown>
-                                                                    {item.description}
-                                                                </ReactMarkdown>
-                                                            </Box>
-                                                        }
-                                                    >
-                                                        <Icon name="status-info" />
-                                                    </Popover>
-                                                ),
-                                            },
-                                            {
-                                                id: "actions",
-                                                header: "Actions",
-                                                cell: (item) => (
-                                                    <Button
-                                                        variant="icon"
-                                                        iconName="close"
-                                                        onClick={() => removeMcpServer(item.name)}
-                                                    />
-                                                ),
-                                            },
-                                        ]}
-                                        items={selectedMcpServersData}
-                                        loadingText="Loading MCP servers"
-                                        empty={
-                                            <Box textAlign="center" color="inherit">
-                                                <b>No MCP servers selected</b>
-                                                <Box
-                                                    padding={{ bottom: "s" }}
-                                                    variant="p"
-                                                    color="inherit"
-                                                >
-                                                    Select MCP servers from the dropdown above.
-                                                </Box>
-                                            </Box>
-                                        }
-                                    />
-                                )}
-                            </Container>
-                        </SpaceBetween>
-                    </Container>
-                </div>
-            ),
-        },
-              ]
-            : []),
-        // Only include Knowledge Bases step if knowledge base is supported and architecture is SINGLE
-        ...(knowledgeBaseIsSupported && architectureType === "SINGLE"
-            ? [
-                  {
-                      title: "Knowledge Bases",
-                      content: (
-                          <div style={{ minHeight: stepMinHeight }}>
-                              <Container
-                                  header={<Header variant="h2">Configure Knowledge Bases</Header>}
-                              >
-                                  <SpaceBetween direction="vertical" size="l">
-                                      <FormField label="Available Knowledge Bases">
-                                          <Select
-                                              placeholder={
-                                                  availableKnowledgeBases.length === 0
-                                                      ? "No additional knowledge bases available"
-                                                      : "Select a knowledge base to add"
-                                              }
-                                              options={availableKnowledgeBases}
-                                              disabled={availableKnowledgeBases.length === 0}
-                                              onChange={({ detail }) => {
-                                                  if (detail.selectedOption) {
-                                                      addKnowledgeBase(detail.selectedOption.value);
-                                                  }
-                                              }}
-                                              selectedOption={null}
-                                          />
-                                      </FormField>
-
-                                      <Container
-                                          header={
-                                              <Header variant="h3">Selected Knowledge Bases</Header>
-                                          }
-                                      >
-                                          {selectedKnowledgeBasesData.length === 0 ? (
-                                              <Alert type="info">No knowledge bases selected</Alert>
-                                          ) : (
-                                              <Table
-                                                  columnDefinitions={[
-                                                      {
-                                                          id: "name",
-                                                          header: "Knowledge Base",
-                                                          cell: (item) => item.name,
-                                                      },
-                                                      {
-                                                          id: "description",
-                                                          header: "Description",
-                                                          cell: (item) => (
-                                                              <Popover
-                                                                  dismissButton={false}
-                                                                  position="top"
-                                                                  size="medium"
-                                                                  triggerType="custom"
-                                                                  content={
-                                                                      <Box padding="xs">
-                                                                          <ReactMarkdown>
-                                                                              {item.description}
-                                                                          </ReactMarkdown>
-                                                                      </Box>
-                                                                  }
-                                                              >
-                                                                  <Icon name="status-info" />
-                                                              </Popover>
-                                                          ),
-                                                      },
-                                                      {
-                                                          id: "parameters",
-                                                          header: "Parameters",
-                                                          cell: (item) => (
-                                                              <Button
-                                                                  variant="normal"
-                                                                  onClick={() =>
-                                                                      openConfigureModal(
-                                                                          item.toolName,
-                                                                      )
-                                                                  }
-                                                              >
-                                                                  Configure
-                                                              </Button>
-                                                          ),
-                                                      },
-                                                      {
-                                                          id: "actions",
-                                                          header: "Actions",
-                                                          cell: (item) => (
-                                                              <Button
-                                                                  variant="icon"
-                                                                  iconName="close"
-                                                                  onClick={() =>
-                                                                      removeTool(item.toolName)
-                                                                  }
-                                                              />
-                                                          ),
-                                                      },
-                                                  ]}
-                                                  items={selectedKnowledgeBasesData}
-                                                  loadingText="Loading knowledge bases"
-                                                  empty={
-                                                      <Box textAlign="center" color="inherit">
-                                                          <b>No knowledge bases selected</b>
-                                                          <Box
-                                                              padding={{ bottom: "s" }}
-                                                              variant="p"
-                                                              color="inherit"
-                                                          >
-                                                              Select knowledge bases from the
-                                                              dropdown above.
-                                                          </Box>
-                                                      </Box>
-                                                  }
-                                              />
-                                          )}
-                                      </Container>
-                                  </SpaceBetween>
-                              </Container>
-                          </div>
-                      ),
-                  },
-              ]
-            : []),
-        ...(architectureType === "SINGLE"
-            ? [
-                  {
-                      title: "Review",
-                      content: (
-                          <div style={{ minHeight: stepMinHeight }}>
-                              <Container
-                                  header={<Header variant="h2">Review Configuration</Header>}
-                              >
-                                  <SpaceBetween direction="vertical" size="m">
-                                      {!isCreating && (
-                                          <Alert type="info" header="Configuration Summary">
-                                              Review your agent configuration before creating.
-                                          </Alert>
-                                      )}
-                                      <Box padding="m" variant="code">
-                                          <pre style={{ margin: 0, overflow: "auto" }}>
-                                              {JSON.stringify(config, null, 2)}
-                                          </pre>
-                                      </Box>
-                                  </SpaceBetween>
-                              </Container>
-                          </div>
-                      ),
-                  },
-              ]
-            : [
-                  {
-                      title: "Swarm Configuration",
-                      content: (
-                          <div style={{ minHeight: stepMinHeight }}>
-                              <SpaceBetween direction="vertical" size="l">
-                                  <Container header={<Header variant="h2">Agent Name</Header>}>
-                                      <FormField
-                                          label="Agent Name"
-                                          description="Enter a unique name for your swarm agent"
-                                          errorText={
-                                              config.agentName.trim() === ""
-                                                  ? "Agent name is required"
-                                                  : !/^[a-zA-Z][a-zA-Z0-9_]{0,47}$/.test(
-                                                          config.agentName,
-                                                      )
-                                                    ? "Agent name must start with a letter and contain only letters, numbers, and underscores (max 48 characters)"
-                                                    : ""
-                                          }
-                                      >
-                                          <Input
-                                              value={config.agentName}
-                                              onChange={({ detail }) =>
-                                                  setConfig((prev) => ({
-                                                      ...prev,
-                                                      agentName: detail.value,
-                                                  }))
-                                              }
-                                              placeholder="Enter agent name..."
-                                              invalid={config.agentName.trim() === ""}
-                                          />
-                                      </FormField>
-                                  </Container>
-
-                                  <Container header={<Header variant="h2">Agent Source</Header>}>
-                                      <SpaceBetween direction="vertical" size="l">
-                                              <SpaceBetween direction="vertical" size="m">
-                                                  <FormField label="Select Agent">
-                                                      <Select
-                                                          placeholder="Select an existing agent to reference"
-                                                          options={availableAgents
-                                                              .filter(
-                                                                  (a) =>
-                                                                      !swarmConfig.agentReferences.some(
-                                                                          (r) =>
-                                                                              r.agentName ===
-                                                                              a.agentName,
-                                                                      ),
-                                                              )
-                                                              .map((a) => ({
-                                                                  label: a.agentName,
-                                                                  value: a.agentName,
-                                                              }))}
-                                                          onChange={({ detail }) => {
-                                                              if (detail.selectedOption?.value) {
-                                                                  addAgentReference(
-                                                                      detail.selectedOption.value,
-                                                                  );
-                                                              }
-                                                          }}
-                                                          selectedOption={null}
-                                                      />
-                                                  </FormField>
-                                                  {swarmConfig.agentReferences.length === 0 ? (
-                                                      <Alert type="info">
-                                                          No agent references added yet.
-                                                      </Alert>
-                                                  ) : (
-                                                      <Table
-                                                          items={swarmConfig.agentReferences}
-                                                          columnDefinitions={[
-                                                              {
-                                                                  id: "agentName",
-                                                                  header: "Agent Name",
-                                                                  cell: (item) => item.agentName,
-                                                                  isRowHeader: true,
-                                                              },
-                                                              {
-                                                                  id: "endpointName",
-                                                                  header: "Endpoint",
-                                                                  cell: (item) => {
-                                                                      const idx = swarmConfig.agentReferences.indexOf(item);
-                                                                      const agent = availableAgents.find(
-                                                                          (a) => a.agentName === item.agentName,
-                                                                      );
-                                                                      const endpointOptions: { label: string; value: string }[] = [];
-                                                                      if (agent?.qualifierToVersion) {
-                                                                          try {
-                                                                              const qtv = JSON.parse(agent.qualifierToVersion);
-                                                                              if (qtv && typeof qtv === "object") {
-                                                                                  endpointOptions.push(
-                                                                                      ...Object.keys(qtv).map((key) => ({
-                                                                                          label: key,
-                                                                                          value: key,
-                                                                                      })),
-                                                                                  );
-                                                                              }
-                                                                          } catch (error) {
-                                                                              console.error("Failed to parse qualifierToVersion:", error);
-                                                                          }
-                                                                      }
-                                                                      if (!endpointOptions.some((o) => o.value === "DEFAULT")) {
-                                                                          endpointOptions.unshift({ label: "DEFAULT", value: "DEFAULT" });
-                                                                      }
-                                                                      return (
-                                                                          <Select
-                                                                              expandToViewport
-                                                                              selectedOption={
-                                                                                  endpointOptions.find(
-                                                                                      (o) => o.value === item.endpointName,
-                                                                                  ) || { label: item.endpointName, value: item.endpointName }
-                                                                              }
-                                                                              onChange={({
-                                                                                  detail,
-                                                                              }) =>
-                                                                                  updateAgentReferenceEndpoint(
-                                                                                      idx,
-                                                                                      detail.selectedOption?.value || "DEFAULT",
-                                                                                  )
-                                                                              }
-                                                                              options={endpointOptions}
-                                                                          />
-                                                                      );
-                                                                  },
-                                                              },
-                                                              {
-                                                                  id: "actions",
-                                                                  header: "Actions",
-                                                                  cell: (item) => {
-                                                                      const idx = swarmConfig.agentReferences.indexOf(item);
-                                                                      return (
-                                                                          <Button
-                                                                              variant="icon"
-                                                                              iconName="close"
-                                                                              onClick={() =>
-                                                                                  removeAgentReference(idx)
-                                                                              }
-                                                                          />
-                                                                      );
-                                                                  },
-                                                              },
-                                                          ]}
-                                                      />
-                                                  )}
-                                              </SpaceBetween>
-                                      </SpaceBetween>
-                                  </Container>
-
-                                  <Container
-                                      header={<Header variant="h2">Entry Agent</Header>}
-                                  >
-                                      <FormField
-                                          label="Entry Agent"
-                                          description="The agent that receives the initial user message"
-                                      >
-                                          <Select
-                                              placeholder="Select entry agent"
-                                              options={getSwarmAgentNames().map((name) => ({
-                                                  label: name,
-                                                  value: name,
-                                              }))}
-                                              selectedOption={
-                                                  swarmConfig.entryAgent
-                                                      ? {
-                                                            label: swarmConfig.entryAgent,
-                                                            value: swarmConfig.entryAgent,
-                                                        }
-                                                      : null
-                                              }
-                                              onChange={({ detail }) =>
-                                                  setSwarmConfig((prev) => ({
-                                                      ...prev,
-                                                      entryAgent:
-                                                          detail.selectedOption?.value || "",
-                                                  }))
-                                              }
-                                              disabled={getSwarmAgentNames().length === 0}
-                                          />
-                                      </FormField>
-                                  </Container>
-
-                                  <Container
-                                      header={
-                                          <Header variant="h2">Orchestrator Settings</Header>
-                                      }
-                                  >
-                                      <ColumnLayout columns={2} variant="text-grid">
-                                          <FormField
-                                              label="Max Handoffs"
-                                              description="Maximum agent-to-agent handoffs"
-                                          >
-                                              <Input
-                                                  type="number"
-                                                  value={swarmConfig.orchestrator.maxHandoffs.toString()}
-                                                  onChange={({ detail }) =>
-                                                      setSwarmConfig((prev) => ({
-                                                          ...prev,
-                                                          orchestrator: {
-                                                              ...prev.orchestrator,
-                                                              maxHandoffs:
-                                                                  parseInt(detail.value) || 15,
-                                                          },
-                                                      }))
-                                                  }
-                                              />
-                                          </FormField>
-                                          <FormField
-                                              label="Max Iterations"
-                                              description="Maximum total iterations"
-                                          >
-                                              <Input
-                                                  type="number"
-                                                  value={swarmConfig.orchestrator.maxIterations.toString()}
-                                                  onChange={({ detail }) =>
-                                                      setSwarmConfig((prev) => ({
-                                                          ...prev,
-                                                          orchestrator: {
-                                                              ...prev.orchestrator,
-                                                              maxIterations:
-                                                                  parseInt(detail.value) || 50,
-                                                          },
-                                                      }))
-                                                  }
-                                              />
-                                          </FormField>
-                                          <FormField
-                                              label="Execution Timeout (s)"
-                                              description="Total execution timeout in seconds"
-                                          >
-                                              <Input
-                                                  type="number"
-                                                  value={swarmConfig.orchestrator.executionTimeoutSeconds.toString()}
-                                                  onChange={({ detail }) =>
-                                                      setSwarmConfig((prev) => ({
-                                                          ...prev,
-                                                          orchestrator: {
-                                                              ...prev.orchestrator,
-                                                              executionTimeoutSeconds:
-                                                                  parseFloat(detail.value) || 300,
-                                                          },
-                                                      }))
-                                                  }
-                                              />
-                                          </FormField>
-                                          <FormField
-                                              label="Node Timeout (s)"
-                                              description="Per-agent timeout in seconds"
-                                          >
-                                              <Input
-                                                  type="number"
-                                                  value={swarmConfig.orchestrator.nodeTimeoutSeconds.toString()}
-                                                  onChange={({ detail }) =>
-                                                      setSwarmConfig((prev) => ({
-                                                          ...prev,
-                                                          orchestrator: {
-                                                              ...prev.orchestrator,
-                                                              nodeTimeoutSeconds:
-                                                                  parseFloat(detail.value) || 60,
-                                                          },
-                                                      }))
-                                                  }
-                                              />
-                                          </FormField>
-                                      </ColumnLayout>
-                                  </Container>
-
-                                  <Container
-                                      header={
-                                          <Header variant="h2">Conversation Manager</Header>
-                                      }
-                                  >
-                                      <FormField label="Conversation Manager">
-                                          <Select
-                                              selectedOption={
-                                                  conversationManagerOptions.find(
-                                                      (opt) =>
-                                                          opt.value ===
-                                                          swarmConfig.conversationManager,
-                                                  ) || null
-                                              }
-                                              onChange={({ detail }) =>
-                                                  setSwarmConfig((prev) => ({
-                                                      ...prev,
-                                                      conversationManager:
-                                                          (detail.selectedOption?.value ||
-                                                              "sliding_window") as
-                                                              | "null"
-                                                              | "sliding_window"
-                                                              | "summarizing",
-                                                  }))
-                                              }
-                                              options={conversationManagerOptions}
-                                          />
-                                      </FormField>
-                                  </Container>
-                              </SpaceBetween>
-                          </div>
-                      ),
-                  },
-                  {
-                      title: "Review",
-                      content: (
-                          <div style={{ minHeight: stepMinHeight }}>
-                              <Container
-                                  header={<Header variant="h2">Review Configuration</Header>}
-                              >
-                                  <SpaceBetween direction="vertical" size="m">
-                                      {!isCreating && (
-                                          <Alert type="info" header="Configuration Summary">
-                                              Review your swarm agent configuration before creating.
-                                          </Alert>
-                                      )}
-                                      <Box padding="m" variant="code">
-                                          <pre style={{ margin: 0, overflow: "auto" }}>
-                                              {JSON.stringify(
-                                                  {
-                                                      agentName: config.agentName,
-                                                      architectureType,
-                                                      swarmConfig,
-                                                  },
-                                                  null,
-                                                  2,
-                                              )}
-                                          </pre>
-                                      </Box>
-                                  </SpaceBetween>
-                              </Container>
-                          </div>
-                      ),
-                  },
-              ]),
-    ];
-
+    // ----------------------------------------------------------------
+    // Wizard shell + configure modal
+    // ----------------------------------------------------------------
     const wizardContent = (
         <Wizard
             i18nStrings={{
@@ -1696,7 +641,6 @@ export default function AgentCoreRuntimeCreatorWizard({
                                     }));
 
                                     if (detail.checked) {
-                                        // Set up default reranking configuration
                                         updateToolParameter(
                                             selectedKbForConfig,
                                             "retrieval_cfg.vectorSearchConfiguration.rerankingConfiguration.type",
@@ -1809,7 +753,6 @@ export default function AgentCoreRuntimeCreatorWizard({
                                     }));
 
                                     if (!detail.checked) {
-                                        // Remove overrideSearchType from config when unchecked
                                         setConfig((prev) => {
                                             const newToolParameters = { ...prev.toolParameters };
                                             const kbParams = {
@@ -1926,7 +869,6 @@ export default function AgentCoreRuntimeCreatorWizard({
                     </SpaceBetween>
                 )}
             </Modal>
-
         </>
     );
 }
